fix: update remaining import paths after refactoring
